Added derivatives from Andrew's "derivatives" branch
