+++ test/runtests.jl
@@ -2,3 +2,4 @@ using Base.Test
 using Rotations
 
 include("rotation_tests.jl")
+#include("derivative_tests.jl")
